Time: 543 ms (12.50%), Space: 30.8 MB (62.50%) - LeetHub

+++ 2147-number-of-ways-to-divide-a-long-corridor/2147-number-of-ways-to-divide-a-long-corridor.cpp
@@ -4,7 +4,7 @@ public:
     int numberOfWays(string corridor) {
         unordered_map<char, int> m, realCount;
         long long int plantCount = 1;
-        long long int curr = 0, mod = 1000000007;
+        int curr = 0, mod = 1000000007;
         for (auto i : corridor) {
             if (m['S'] == 2) {
                 if (i == 'P') curr = (curr + 1) % mod;
